Adds lizards to roundstart races.
I also put them in the main pool, because I didnt want to figure out where "join_with_muntant_race" was located, and having to worry about it being turned off. I'll add in the no legion lizard and like the code later, because I need to actually test that it works.